Rename ast
Prompt: "Rename ast.py to code.py"
